Hack?  Or workaround?
This once again allows a reboot to work; the problem appears
to be that launchd isn't being notified of a service being shut
down, and so it just loops infinitely.  Which is dumb, since the
process by this point has been sent a SIGKILL, and if it's not already
dead, it's never going to die.

(cherry picked from commit 98cb617e2232cf516eee858cc6d1c00fb7acd8b5)